move flowchartlabel in fragments directory
